feature: renew authen page

+++ client/src/api/errorHandler.js
@@ -1,6 +1,7 @@
 export default function errorHandler(e) {
+  if (!e.response) return
   switch (e.response.status) {
-    case 401:
+    case 403:
       localStorage.clear()
       window.location.href = '/login'
       break
